inference/requirements (2).txt
boto3
requests
google-generativeai
fastapi
pydantic
diffusers
huggingface_hub
Pillow
realesrgan
transformers 
accelerate 
scipy 
huggingface_hub 
peft
sentencepiece
uvicorn
basicsr
facexlib
gfpgan
smart_open